requirements.txt
flask
python-dotenv
flask-restful
flask_sqlalchemy==2.5.1
flask-jwt-extended
flask-mail
SQLAlchemy==1.4.23
